backend/requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
rank-bm25
numpy
pydantic
httpx
